tests/wpt/web-platform-tests/navigation-api/navigation-methods/return-value/navigate-push-not-loaded.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/helpers.js"></script>

<script>
promise_test(async t => {
  // Purposefully do not wait until after the load event (unlike some sibling tests).

  navigation.onnavigate = t.unreached_func("onnavigate should not be called");
  navigation.onnavigatesuccess = t.unreached_func("onnavigatesuccess should not be called");
  navigation.onnavigateerror = t.unreached_func("onnavigateerror should not be called");

  assert_equals(document.readyState, "loading", "Document must not have loaded yet");

  const result = navigation.navigate("#1", { history: "push" });
  await assertBothRejectDOM(t, result, "NotSupportedError");
}, "navigate() with history: 'push' in a document that has not yet had its load event");
</script>
